# apps/backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "VRK API",
        "contact": {}
    },
    "basePath": "/api/v1",
    "paths": {
        "/agreements": {
            "get": {
                "description": "Returns contracts visible to the authenticated customer or contractor organization through the agreements adapter boundary.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agreements"
                ],
                "summary": "List contracts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/agreement.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/agreement.AgreementResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/agreement.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/agreement.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a Stage 03 contract bound to a customer organization and contractor organization while preserving the backend agreements adapter boundary.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agreements"
                ],
                "summary": "Create contract",
                "parameters": [
                    {
                        "description": "Contract payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/agreement.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/agreement.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/agreement.AgreementResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/agreement.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/agreement.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/agreement.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/agreement.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/agreement.Response"
                        }
                    }
                }
            }
        },
        "/agreements/contractors": {
            "get": {
                "description": "Returns active contractor organizations that can be bound to a customer contract.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agreements"
                ],
                "summary": "List active contractor organizations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/agreement.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/agreement.ContractorOptionResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/agreement.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/agreement.Response"
                        }
                    }
                }
            }
        },
        "/agreements/routing/resolve": {
            "post": {
                "description": "Resolves routing-eligible contracts and the allowed contractor from customer contract context without enabling live Stage 04 request creation.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agreements"
                ],
                "summary": "Resolve contract routing",
                "parameters": [
                    {
                        "description": "Routing resolve payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/agreement.RoutingResolveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/agreement.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/agreement.RoutingResolveResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/agreement.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/agreement.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/agreement.Response"
                        }
                    }
                }
            }
        },
        "/agreements/{id}": {
            "get": {
                "description": "Returns one contract visible to the authenticated customer or contractor organization.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agreements"
                ],
                "summary": "Get contract by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Contract ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/agreement.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/agreement.AgreementResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/agreement.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/agreement.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/agreement.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates one customer-owned contract through the legacy agreements path while keeping the public web contour named contracts.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agreements"
                ],
                "summary": "Update contract",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Contract ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Contract update payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/agreement.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/agreement.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/agreement.AgreementResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/agreement.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/agreement.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/agreement.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/agreement.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/agreement.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "agreement.AgreementLocationScopeResponse": {
            "type": "object",
            "properties": {
                "label": {
                    "type": "string"
                },
                "scopeId": {
                    "type": "string"
                },
                "scopeType": {
                    "type": "string"
                }
            }
        },
        "agreement.AgreementResponse": {
            "type": "object",
            "properties": {
                "contractNumber": {
                    "type": "string"
                },
                "contractStatus": {
                    "type": "string"
                },
                "contractorOrganizationId": {
                    "type": "string"
                },
                "contractorOrganizationName": {
                    "type": "string"
                },
                "customerOrganizationId": {
                    "type": "string"
                },
                "customerOrganizationName": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "equipmentType": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "locationScope": {
                    "$ref": "#/definitions/agreement.AgreementLocationScopeResponse"
                },
                "region": {
                    "type": "string"
                },
                "routingEligible": {
                    "type": "boolean"
                },
                "source": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                },
                "subjectOfAgreement": {
                    "type": "string"
                },
                "workType": {
                    "type": "string"
                }
            }
        },
        "agreement.ContractorOptionResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "shortName": {
                    "type": "string"
                }
            }
        },
        "agreement.CreateRequest": {
            "type": "object",
            "properties": {
                "contractNumber": {
                    "type": "string"
                },
                "contractStatus": {
                    "type": "string"
                },
                "contractorOrganizationId": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "equipmentType": {
                    "type": "string"
                },
                "locationScopeLabel": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                },
                "subdivisionId": {
                    "type": "string"
                },
                "subjectOfAgreement": {
                    "type": "string"
                },
                "unitId": {
                    "type": "string"
                },
                "workType": {
                    "type": "string"
                }
            }
        },
        "agreement.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "agreement.RoutingResolutionItem": {
            "type": "object",
            "properties": {
                "contract": {
                    "$ref": "#/definitions/agreement.AgreementResponse"
                },
                "contractor": {
                    "$ref": "#/definitions/agreement.ContractorOptionResponse"
                }
            }
        },
        "agreement.RoutingResolveRequest": {
            "type": "object",
            "properties": {
                "equipmentType": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "unitId": {
                    "type": "string"
                },
                "workType": {
                    "type": "string"
                }
            }
        },
        "agreement.RoutingResolveResponse": {
            "type": "object",
            "properties": {
                "equipmentType": {
                    "type": "string"
                },
                "matches": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/agreement.RoutingResolutionItem"
                    }
                },
                "region": {
                    "type": "string"
                },
                "unitId": {
                    "type": "string"
                },
                "workType": {
                    "type": "string"
                }
            }
        },
        "agreement.UpdateRequest": {
            "type": "object",
            "properties": {
                "contractNumber": {
                    "type": "string"
                },
                "contractStatus": {
                    "type": "string"
                },
                "contractorOrganizationId": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "equipmentType": {
                    "type": "string"
                },
                "locationScopeLabel": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                },
                "subdivisionId": {
                    "type": "string"
                },
                "subjectOfAgreement": {
                    "type": "string"
                },
                "unitId": {
                    "type": "string"
                },
                "workType": {
                    "type": "string"
                }
            }
        },
...
    }
}